Seperate Date and Tags components to enhance code organization and reusability (SRP)
